ml_backend/requirements.txt.txt
fastapi
uvicorn
python-multipart
pillow
torch
ultralytics